+++ identity-wallet/src/state/credentials/reducers/handle_credential_offer.rs
