+++ identity-wallet/src/state/credentials/reducers/handle_oid4vp_authorization_request.rs
@@ -464,7 +464,7 @@ async fn get_vp_token(
                 // "Which DID/key do we sign with?" is answered differently at every call site: the `cnf`
                 // claim here, from `credentialSubject.id` when signing a DACT, and from the preferred
                 // DID method elsewhere. Each answer has to match the DID the credential was issued to, so they
-                // all need to behave identically. This should be unified into a single helper. 
+                // all need to behave identically. This should be unified into a single helper.
                 let Some(RequiredKeyBinding::Kid(cnf_kid)) = vcdm2_sd_jwt.claims().cnf.as_ref() else {
                     return Err(AppError::Error(
                         "Unsupported `cnf` claim in VCDM 2.0 SD-JWT".to_string(),
