+++ identity-wallet/src/state/credentials/reducers/mod.rs
@@ -1,9 +1,9 @@
 pub mod delete_credential;
+pub mod handle_credential_offer;
 pub mod handle_oid4vp_authorization_request;
 pub mod refresh_all_credential_statuses;
 pub mod refresh_credential_status;
 pub mod self_issue_credential;
-pub mod send_credential_request;
 pub mod send_interactive_authorization_request_follow_up;
 pub mod send_token_request;
 pub mod share_to_linkedin;
