+++ identity-wallet/src/state/qr_code/reducers/mod.rs
@@ -1,3 +1,3 @@
 pub mod accept_connection;
-pub mod read_authorization_request;
 pub mod read_credential_offer;
+pub mod read_oid4vp_authorization_request;
