+++ identity-wallet/src/state/qr_code/reducers/read_oid4vp_authorization_request.rs
