+++ unime/src-tauri/tests/fixtures/states/four_credentials_redirect_me.json
@@ -151,7 +151,7 @@
       "display_name": "PersonalInformation"
     },
     {
-      "id": "65323136-6535-3737-6463-386531323361",
+      "id": "33386234-3632-3438-3932-313462316336",
       "issuer_name": "",
       "format": {
         "format": "jwt_vc_json"
@@ -259,7 +259,7 @@
         {
           "title": "Teamwork",
           "issuer_name": "Impierce Demo Portal",
-          "id": "65323136-6535-3737-6463-386531323361"
+          "id": "33386234-3632-3438-3932-313462316336"
         }
       ]
     },
@@ -272,7 +272,7 @@
         {
           "title": "Teamwork",
           "issuer_name": "Impierce Demo Portal",
-          "id": "65323136-6535-3737-6463-386531323361"
+          "id": "33386234-3632-3438-3932-313462316336"
         }
       ]
     }
